Migrate to maven, optimize code and world optimization
