--- nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=158e1339
-build.xml.script.CRC32=f18199eb
-build.xml.stylesheet.CRC32=8064a381@1.79.1.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=158e1339
-nbproject/build-impl.xml.script.CRC32=3f386506
-nbproject/build-impl.xml.stylesheet.CRC32=05530350@1.79.1.48
